fix(nats): typed validation guards + structured exceptions + unified pattern
- Add MalformedRequestError(ValueError) for explicit domain-level exception
- SttRequest/TtsRequest.from_payload: reject bool for numeric fields,
  add isinstance(str) guards for optional string fields, reject non-string
  request_id before regex match
- STT adapter: delegate handle() to validate_stt_request() to unify
  with TTS delegated pattern (fixes parallel-path-drift)
- Extract SttRequest + TtsRequest to voicecli.adapters.nats.requests
  to break circular dependency with _validation.py and keep file
  length under 300 lines
- _validation.py: module-level imports from .requests, catch
  MalformedRequestError instead of broad ValueError
